Hostapd: set service to default disable
The hostapd package is included in the system and for  some users that will run
other wlan modes cant do it if the system is already
started hostapd. This hostapd bbappend is an example of
how to add a AP mode network. You can try it out
with systemctl start hostapd. this example will add a
network ssid HostMobility-Test-Network.