ntp: Replace ntp servers
Replace pool.ntp.org in NTPSERVERS with time servers from Cloudflare
and NIST. The TOS for pool.ntp.org state that one should to register
as a vendor and "must absolutely not use the default pool.ntp.org zone
names as the default configuration in your application or appliance."

Also, stop including any more servers from /etc/ntp.servers. If you
are unhappy with the servers in this list, simply modify this file in
your custom distro.