Add hm-generic-console distro, this olny require  yocto-uninative.inc and security_flags.inc
This uses the same distro version as poky does but does not include all distro features that poky do.
Use this distro when you only need console and no gui/desktop version.
hmm and fgeneric is now default this distro instead of poky. can be override with enviorment export DISTRO.